Move misc classes with inner classes to their own packages
